pyproject.toml
[project]
name = "factorypattern"
version = "0.1.0"
description = "Factory Pattern Tutorial with Python ABC, FastAPI and different repository implementations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.129.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
